+++ pkgs/standards/auto_authn/auto_authn/oidc_userinfo.py
@@ -23,7 +23,7 @@ from fastapi import (
 
 from .fastapi_deps import get_current_principal
 from .jwtoken import JWTCoder, InvalidTokenError, _svc
-from .orm.tables import User
+from .orm import User
 from .rfc6750 import extract_bearer_token
 from .deps import JWAAlg
