+++ pkgs/standards/auto_authn/tests/unit/test_authorize_id_token_hashes.py
@@ -5,7 +5,7 @@ import pytest
 from fastapi import status
 
 from auto_authn.crypto import hash_pw
-from auto_authn.orm.tables import Client, Tenant, User
+from auto_authn.orm import Client, Tenant, User
 from auto_authn.oidc_id_token import oidc_hash, verify_id_token
 from auto_authn.rfc8414_metadata import ISSUER
 from auto_authn.routers.auth_flows import AUTH_CODES
