+++ pkgs/standards/auto_authn/tests/unit/test_authorize_response_modes.py
@@ -5,7 +5,7 @@ import pytest
 from fastapi import status
 
 from auto_authn.crypto import hash_pw
-from auto_authn.orm.tables import Client, Tenant, User
+from auto_authn.orm import Client, Tenant, User
 
 
 async def _setup(async_client, db_session):
