+++ pkgs/standards/auto_authn/tests/unit/test_models.py
@@ -1,5 +1,5 @@
 """
-Unit tests for auto_authn.orm.tables module.
+Unit tests for auto_authn.orm module.
 
 Tests ORM model creation, validation, relationships, hybrid properties, and methods.
 """
@@ -8,7 +8,7 @@ import uuid
 
 import pytest
 
-from auto_authn.orm.tables import Tenant, Client, User, Service, ApiKey, ServiceKey
+from auto_authn.orm import Tenant, Client, User, Service, ApiKey, ServiceKey
 from auto_authn.crypto import hash_pw
 
 
@@ -192,7 +192,7 @@ class TestModelIntegration:
 
     def test_client_id_regex_pattern(self):
         """Test the client ID regex pattern directly."""
-        from auto_authn.orm.tables import _CLIENT_ID_RE
+        from auto_authn.orm import _CLIENT_ID_RE
 
         valid_patterns = [
             "simple123",
