+++ pkgs/standards/auto_authn/tests/unit/test_oidc_authorize_scope_nonce.py
@@ -7,7 +7,7 @@ import pytest
 from fastapi import status
 
 from auto_authn.crypto import hash_pw
-from auto_authn.orm.tables import Client, Tenant, User
+from auto_authn.orm import Client, Tenant, User
 from auto_authn.oidc_discovery import ISSUER
 from auto_authn.oidc_id_token import verify_id_token
 from auto_authn.routers.auth_flows import AUTH_CODES, SESSIONS
